list of orders handled

+++ src/screens/Restaurant.js
@@ -5,11 +5,8 @@ import '../styles/font/flaticon.css'
 import '../styles/main.css'
 import Navbar from '../components/Navbar';
 import Header from '../components/Header';
-import OrderRow from '../components/OrderRow'
 import Footer from '../components/Footer';
 import toPersianNum from '../utils/PersianNumber';
-import { Redirect } from 'react-router';
-import RestaurantImg from '../images/McDonald\'sLogo.png'
 import axios from 'axios'
 
 import FoodCart from '../components/FoodCart';
@@ -25,6 +22,7 @@ class Restaurant extends Component {
         this.increaseCurrentFood = this.increaseCurrentFood.bind(this)
         this.decreaseCurrentFood = this.decreaseCurrentFood.bind(this)
         this.addFoodFromModal = this.addFoodFromModal.bind(this)
+        this.finalizeOrder = this.finalizeOrder.bind(this)
 
         this.state = {
             loading: true,
@@ -124,6 +122,9 @@ class Restaurant extends Component {
     decreaseFood(index) {
         let tempOrder = this.state.currentOrder.slice()
         tempOrder[index].count -= 1
+        if (tempOrder[index].count < 0) {
+            return;
+        }
         this.setState({currentOrder: tempOrder})
         
         axios.delete('http://localhost:8080/food/' + this.state.restaurantId, {params: {
@@ -156,6 +157,17 @@ class Restaurant extends Component {
         .catch((error) => {
             console.log(error);
         });
+        this.hideFoodModal();
+    }
+    finalizeOrder() {
+        axios.put('http://localhost:8080/order')
+        .then((response) => {
+            this.setState({currentOrder: null})
+        })
+        .catch((error) => {
+            console.log(error);
+        });
+        this.forceUpdate();
     }
 
     render() {
